requirements.txt
paho-mqtt~=2.1.0
dash~=2.18.2
pandas~=2.2.3
plotly~=5.24.1